Remote Thread tracking:
 - Track which thread is used to build a job on remote hosts. This will be used for further changes to improve build visualization
 - Change ThreadIndex to uint16_t to avoid any increase in Job class size

+++ Code/Tools/FBuild/FBuildCore/WorkerPool/WorkerThread.cpp
@@ -22,12 +22,12 @@
 
 // Static
 //------------------------------------------------------------------------------
-static THREAD_LOCAL uint32_t s_WorkerThreadThreadIndex = 0;
+static THREAD_LOCAL uint16_t s_WorkerThreadThreadIndex = 0;
 Mutex WorkerThread::s_TmpRootMutex;
 AStackString<> WorkerThread::s_TmpRoot;
 
 //------------------------------------------------------------------------------
-WorkerThread::WorkerThread( uint32_t threadIndex )
+WorkerThread::WorkerThread( uint16_t threadIndex )
 : m_ShouldExit( false )
 , m_Exited( false )
 , m_ThreadIndex( threadIndex )
@@ -105,7 +105,7 @@ void WorkerThread::WaitForStop()
 
 // GetThreadIndex
 //------------------------------------------------------------------------------
-/*static*/ uint32_t WorkerThread::GetThreadIndex()
+/*static*/ uint16_t WorkerThread::GetThreadIndex()
 {
     return s_WorkerThreadThreadIndex;
 }
